+++ plutus-pab/plutus-pab.cabal
@@ -78,7 +78,6 @@ library
         Plutus.PAB.Core.ContractInstance.BlockchainEnv
         Plutus.PAB.Core.ContractInstance.RequestHandlers
         Plutus.PAB.Core.ContractInstance.STM
-        Plutus.PAB.Core.Server
         Plutus.PAB.Db.Eventful.ContractDefinitionStore
         Plutus.PAB.Db.Eventful.ContractStore
         Plutus.PAB.Db.Eventful.Projections
