+++ plutus-pab/src/Plutus/PAB/Webserver/Server.hs
@@ -12,99 +12,93 @@
 {-# OPTIONS_GHC -fno-warn-partial-type-signatures #-}
 
 module Plutus.PAB.Webserver.Server
-    ( main
+    ( startServer
     ) where
 
-import qualified Cardano.BM.Configuration.Model                 as CM
-import           Cardano.BM.Trace                               (Trace)
-import           Cardano.Node.Types                             (MockServerConfig (..))
-import           Control.Concurrent.Availability                (Availability, available)
-import qualified Control.Concurrent.STM                         as STM
-import           Control.Monad                                  ((<=<))
-import           Control.Monad.Except                           (ExceptT (ExceptT))
-import           Control.Monad.Freer                            (Eff, interpret)
-import           Control.Monad.Freer.Reader                     (runReader)
-import           Control.Monad.IO.Class                         (liftIO)
-import           Data.Bifunctor                                 (first)
-import qualified Data.ByteString.Lazy.Char8                     as LBS
-import           Data.Function                                  ((&))
-import           Data.Proxy                                     (Proxy (Proxy))
-import qualified Data.Text.Encoding                             as Text
-import qualified Network.Wai.Handler.Warp                       as Warp
-import           Servant                                        (Application, Handler (Handler), Raw, ServerT, err400,
-                                                                 err500, errBody, hoistServer, serve,
-                                                                 serveDirectoryFileServer, (:<|>) ((:<|>)))
-import           Servant.Client                                 (BaseUrl (baseUrlPort))
+import           Control.Concurrent              (forkIO)
+import           Control.Concurrent.Availability (Availability, available)
+import qualified Control.Concurrent.STM          as STM
+import           Control.Monad                   (void)
+import           Control.Monad.Except            (ExceptT (ExceptT))
+import           Control.Monad.IO.Class          (liftIO)
+import           Data.Aeson                      (FromJSON, ToJSON)
+import           Data.Bifunctor                  (first)
+import qualified Data.ByteString.Lazy.Char8      as LBS
+import           Data.Function                   ((&))
+import           Data.Proxy                      (Proxy (Proxy))
+import qualified Network.Wai.Handler.Warp        as Warp
+import           Servant                         (Application, Handler (Handler), Raw, ServerT, err500, errBody,
+                                                  hoistServer, serve, serveDirectoryFileServer, (:<|>) ((:<|>)))
+import           Servant.Client                  (BaseUrl (baseUrlPort))
 
 
-import           Control.Monad.Freer.Extras.Log                 (LogMsg, logInfo, mapLog)
-import           Plutus.PAB.App                                 (App, AppBackend, runApp)
-import           Plutus.PAB.Arbitrary                           ()
-import           Plutus.PAB.Core.ContractInstance               (ContractInstanceMsg)
-import qualified Plutus.PAB.Core.ContractInstance.BlockchainEnv as BlockchainEnv
-import           Plutus.PAB.Core.ContractInstance.STM           (BlockchainEnv, InstancesState)
-import qualified Plutus.PAB.Core.ContractInstance.STM           as ContractInstance
-import           Plutus.PAB.Effects.Contract.ContractExe        (ContractExe)
-import qualified Plutus.PAB.Effects.Contract.ContractExe        as LM
-import qualified Plutus.PAB.Monitoring.PABLogMsg                as LM
-import           Plutus.PAB.ParseStringifiedJSON                (UnStringifyJSONLog)
-import           Plutus.PAB.Types                               (Config (..), PABError (InvalidUUIDError), baseUrl,
-                                                                 pabWebserverConfig, staticDir)
-import           Plutus.PAB.Webserver.API                       (API, WSAPI)
-import           Plutus.PAB.Webserver.Handler                   (handler)
-import           Plutus.PAB.Webserver.Types                     (WebSocketLogMsg)
-import           Plutus.PAB.Webserver.WebSocket                 (handleWS)
-import qualified Wallet.Emulator.Wallet                         as Wallet
+import           Control.Monad.Freer.Extras.Log  (logInfo)
+import           Plutus.PAB.Core                 (PABAction, PABRunner (..))
+import qualified Plutus.PAB.Core                 as Core
+import qualified Plutus.PAB.Effects.Contract     as Contract
+import qualified Plutus.PAB.Monitoring.PABLogMsg as LM
+import           Plutus.PAB.Types                (PABError, WebserverConfig (..), baseUrl)
+import           Plutus.PAB.Webserver.API        (API, NewAPI, WSAPI)
+import           Plutus.PAB.Webserver.Handler    (handlerNew, handlerOld)
+import qualified Plutus.PAB.Webserver.WebSocket  as WS
+import qualified Servant
 
-runHandler :: InstancesState -> Trace IO LM.PABLogMsg -> CM.Configuration -> Config -> Eff (LogMsg LM.ContractExeLogMsg ': LogMsg (ContractInstanceMsg ContractExe) ': LogMsg WebSocketLogMsg ': LogMsg UnStringifyJSONLog ': AppBackend _) a -> IO (Either PABError a)
-runHandler instancesState trace logConfig config =
-  runApp instancesState trace logConfig config
-      . interpret (mapLog LM.SUnstringifyJSON)
-      . interpret (mapLog LM.SWebsocketMsg)
-      . interpret (mapLog LM.SContractInstanceMsg)
-      . interpret (mapLog LM.SContractExeLogMsg)
+asHandler :: forall t env a. PABRunner t env -> PABAction t env a -> Handler a
+asHandler PABRunner{runPABAction} = Servant.Handler . ExceptT . fmap (first mapError) . runPABAction where
+    mapError :: PABError -> Servant.ServerError
+    mapError e = Servant.err500 { Servant.errBody = LBS.pack $ show e }
 
-asHandler :: InstancesState -> Trace IO LM.PABLogMsg -> CM.Configuration -> Config -> Eff (LogMsg LM.ContractExeLogMsg ': LogMsg (ContractInstanceMsg ContractExe) ': LogMsg WebSocketLogMsg ': LogMsg UnStringifyJSONLog ': AppBackend _) a -> Handler a
-asHandler instancesState trace logConfig config =
-  Handler . ExceptT . fmap (first decodeErr) . runHandler instancesState trace logConfig config
-  where
-    decodeErr (InvalidUUIDError t) =
-        err400
-            {errBody = "Invalid UUID: " <> LBS.fromStrict (Text.encodeUtf8 t)}
-    decodeErr err = err500 {errBody = LBS.pack $ show err}
+type CombinedAPI t =
+      API (Contract.ContractDef t)
+      :<|> WSAPI
+      :<|> NewAPI (Contract.ContractDef t)
+      -- :<|> Raw
 
-app :: InstancesState -> BlockchainEnv -> Trace IO LM.PABLogMsg -> CM.Configuration -> Config -> Application
-app instancesState env trace logConfig config = do
-  let
-    hdl :: forall a. Eff _ a -> IO a
-    hdl = either (error . show) pure <=< runHandler instancesState trace logConfig config
-            . interpret (mapLog (LM.SWalletEvent . Wallet.RequestHandlerLog))
-            . interpret (mapLog (LM.SWalletEvent . Wallet.TxBalanceLog))
-            . runReader env
+app ::
+    forall t env.
+    ( FromJSON (Contract.ContractDef t)
+    , ToJSON (Contract.ContractDef t)
+    , Contract.PABContract t
+    , Servant.MimeUnrender Servant.JSON (Contract.ContractDef t)
+    ) => WebserverConfig -> PABRunner t env -> Application
+app WebserverConfig{staticDir} pabRunner = do
+    let rest = Proxy @(CombinedAPI t :<|> Raw)
+        apiServer :: ServerT (CombinedAPI t) Handler
+        apiServer =
+            Servant.hoistServer
+                (Proxy @(CombinedAPI t))
+                (asHandler pabRunner)
+                (handlerOld :<|> WS.combinedWebsocket :<|> handlerNew) -- :<|> fileServer)
 
-    rest = Proxy @((API ContractExe :<|> WSAPI) :<|> Raw)
-    apiServer :: ServerT (API ContractExe :<|> WSAPI) Handler
-    apiServer =
-      hoistServer
-        (Proxy @(API ContractExe :<|> WSAPI))
-        (asHandler instancesState trace logConfig config)
-        (handler hdl :<|> handleWS instancesState trace logConfig)
+        fileServer :: ServerT Raw Handler
+        fileServer = serveDirectoryFileServer staticDir
 
-    fileServer :: ServerT Raw Handler
-    fileServer = serveDirectoryFileServer (staticDir . pabWebserverConfig $ config)
-  serve rest (apiServer :<|> fileServer)
+    Servant.serve rest (apiServer :<|> fileServer)
 
 
-main :: Trace IO LM.PABLogMsg -> CM.Configuration -> Config -> Availability -> App ()
-main trace logConfig config availability = interpret (mapLog LM.SContractExeLogMsg) $ do
+-- | Start the server using the config. Returns an action that shuts it down again.
+startServer ::
+    forall t env.
+    ( FromJSON (Contract.ContractDef t)
+    , ToJSON (Contract.ContractDef t)
+    , Contract.PABContract t
+    , Servant.MimeUnrender Servant.JSON (Contract.ContractDef t)
+    ) => WebserverConfig -> Availability -> PABAction t env (PABAction t env ())
+startServer config@WebserverConfig{baseUrl} availability = do
+    let port = baseUrlPort baseUrl
+    simRunner <- Core.pabRunner
+    shutdownVar <- liftIO $ STM.atomically $ STM.newEmptyTMVar @()
 
-    let Config{nodeServerConfig=MockServerConfig{mscSocketPath}} = config
-    let port = baseUrlPort $ baseUrl $ pabWebserverConfig config
-    let warpSettings :: Warp.Settings
+    let shutdownHandler :: IO () -> IO ()
+        shutdownHandler doShutdown = void $ forkIO $ do
+            STM.atomically $ STM.takeTMVar shutdownVar
+            putStrLn "webserver: shutting down"
+            doShutdown
+        warpSettings :: Warp.Settings
         warpSettings = Warp.defaultSettings
             & Warp.setPort port
+            & Warp.setInstallShutdownHandler shutdownHandler
             & Warp.setBeforeMainLoop (available availability)
-    logInfo $ LM.StartingPABBackendServer port
-    instancesState <- liftIO $ STM.atomically $ ContractInstance.emptyInstancesState
-    env <- liftIO $ BlockchainEnv.startNodeClient mscSocketPath instancesState
-    liftIO $ Warp.runSettings warpSettings $ app instancesState env trace logConfig config
+    logInfo @(LM.PABMultiAgentMsg t) (LM.StartingPABBackendServer port)
+    _ <- liftIO $ forkIO $ Warp.runSettings warpSettings $ app config simRunner
+    pure (liftIO $ STM.atomically $ STM.putTMVar shutdownVar ())
