+++ plutus-pab/src/Plutus/PAB/Webserver/WebSocket.hs
@@ -2,137 +2,114 @@
 {-# LANGUAGE FlexibleContexts    #-}
 {-# LANGUAGE GADTs               #-}
 {-# LANGUAGE KindSignatures      #-}
+{-# LANGUAGE NamedFieldPuns      #-}
 {-# LANGUAGE OverloadedStrings   #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications    #-}
 
 module Plutus.PAB.Webserver.WebSocket
-    ( handleWS
+    ( combinedWebsocket
+    , contractInstanceUpdates
+    -- * Reports
+    , getContractReport
     ) where
 
-import qualified Cardano.BM.Configuration.Model          as CM
-import           Cardano.BM.Trace                        (Trace)
-import           Control.Concurrent.Async                (Async, async, waitAnyCancel)
-import           Control.Exception                       (SomeException, handle)
-import           Control.Monad                           (void, when)
-import           Control.Monad.Freer                     (Eff, LastMember, Member)
-import           Control.Monad.Freer.Delay               (DelayEffect, delayThread, handleDelayEffect)
-import           Control.Monad.Freer.Extras.Log          (LogMsg, logInfo)
-import           Control.Monad.Freer.Reader              (Reader, ask)
-import           Control.Monad.Freer.WebSocket           (WebSocketEffect, acceptConnection, sendJSON)
-import           Control.Monad.IO.Class                  (MonadIO, liftIO)
-import           Data.Aeson                              (ToJSON)
-import           Data.Time.Units                         (Second, TimeUnit)
-import           Network.WebSockets.Connection           (Connection, PendingConnection, withPingThread)
-import           Plutus.PAB.App                          (runApp)
-import           Plutus.PAB.Core.ContractInstance.STM    (InstancesState)
-import           Plutus.PAB.Effects.Contract             (ContractDefinitionStore, ContractEffect, ContractStore)
-import           Plutus.PAB.Effects.Contract.ContractExe (ContractExe)
-import qualified Plutus.PAB.Monitoring.PABLogMsg         as LM
-import           Plutus.PAB.Types                        (Config, PABError)
-import           Plutus.PAB.Webserver.Handler            (getChainReport, getContractReport)
-import           Plutus.PAB.Webserver.Types              (StreamToClient (NewChainReport, NewContractReport),
-                                                          WebSocketLogMsg (ClosedConnection, CreatedConnection))
-import           Wallet.Effects                          (ChainIndexEffect)
+import           Control.Applicative                  (Alternative (..))
+import qualified Control.Concurrent.STM               as STM
+import           Control.Exception                    (SomeException, handle)
+import           Control.Monad                        (guard)
+import           Control.Monad.IO.Class               (liftIO)
+import qualified Data.Aeson                           as JSON
+import           Data.Map                             as Map
+import           Data.Proxy                           (Proxy (..))
+import qualified Network.WebSockets                   as WS
+import           Network.WebSockets.Connection        (Connection, PendingConnection)
+import           Plutus.PAB.Core                      (PABAction)
+import qualified Plutus.PAB.Core                      as Core
+import           Plutus.PAB.Core.ContractInstance.STM (OpenEndpoint (..))
+import qualified Plutus.PAB.Effects.Contract          as Contract
+import           Plutus.PAB.Webserver.API             (StatusStreamToClient (..))
+import           Plutus.PAB.Webserver.Types           (ContractReport (..), ContractSignatureResponse (..))
+import           Wallet.Types                         (ContractInstanceId (..))
 
 ------------------------------------------------------------
 -- Message processors.
 ------------------------------------------------------------
-chainReportThread ::
-       ( Member ChainIndexEffect effs
-       , Member DelayEffect effs
-       , Member WebSocketEffect effs
-       )
-    => Connection
-    -> Eff effs ()
-chainReportThread = watchAndNotify (5 :: Second) getChainReport NewChainReport
 
-contractStateThread ::
-       ( Member WebSocketEffect effs
-       , Member (ContractEffect ContractExe) effs
-       , Member DelayEffect effs
-       , Member (ContractDefinitionStore ContractExe) effs
-       , Member (ContractStore ContractExe) effs
-       )
-    => Connection
-    -> Eff effs ()
-contractStateThread =
-    watchAndNotify (3 :: Second) (getContractReport @ContractExe) NewContractReport
+-- contractStateThread ::
+--        ( Member WebSocketEffect effs
+--        , Member (ContractEffect ContractExe) effs
+--        , Member DelayEffect effs
+--        , Member (ContractDefinitionStore ContractExe) effs
+--        , Member (ContractStore ContractExe) effs
+--        )
+--     => Connection
+--     -> Eff effs ()
+-- contractStateThread _ = pure () -- FIXME
+--     -- watchAndNotify (3 :: Second) (getContractReport @ContractExe) NewContractReport
 
-watchAndNotify ::
-       ( TimeUnit t
-       , Member DelayEffect effs
-       , Member WebSocketEffect effs
-       , Eq a
-       , ToJSON b
-       )
-    => t
-    -> Eff effs a
-    -> (a -> b)
-    -> Connection
-    -> Eff effs ()
-watchAndNotify time query wrapper connection =
-    watchForChanges time query (sendJSON connection . wrapper)
+-- watchAndNotify ::
+--        ( TimeUnit t
+--        , Member DelayEffect effs
+--        , Member WebSocketEffect effs
+--        , Eq a
+--        , ToJSON b
+--        )
+--     => t
+--     -> Eff effs a
+--     -> (a -> b)
+--     -> Connection
+--     -> Eff effs ()
+-- watchAndNotify time query wrapper connection = pure () -- FIXME
+--     -- watchForChanges time query (sendJSON connection . wrapper)
 
--- TODO Polling is icky. But we can't use Eventful's hook system
--- because that relies all events coming in from the same thread. We
--- can't guarantee that. (eg. We could be running on the server, but
--- the update came via the CLI.)
---
--- Can we use the DB commit hook instead?
--- https://www.sqlite.org/c3ref/commit_hook.html
-watchForChanges ::
-       (TimeUnit t, Eq a, Member DelayEffect effs)
-    => t
-    -> Eff effs a
-    -> (a -> Eff effs ())
-    -> Eff effs ()
-watchForChanges time query notify = go Nothing
-  where
-    go oldValue = do
-        newValue <- query
-        when (oldValue /= Just newValue) (notify newValue)
-        delayThread time
-        go $ Just newValue
+combinedWebsocket :: forall t env. PendingConnection -> PABAction t env ()
+combinedWebsocket _ = do
+    -- Core.PABRunner{Core.runPABAction} <- Core.pabRunner
+    -- Subscribe / unsubscribe to updates
+    -- slot updates
+    pure ()
 
-------------------------------------------------------------
--- Plumbing
-------------------------------------------------------------
-threadApp :: InstancesState -> Trace IO LM.PABLogMsg -> CM.Configuration -> Config -> Connection -> IO ()
-threadApp instancesState trace logConfig config connection = do
-    tasks :: [Async (Either PABError ())] <-
-        traverse
-            asyncApp
-            [ chainReportThread connection
-            , contractStateThread connection
-            ]
-    void $ waitAnyCancel tasks
-  where
-    asyncApp = async . runApp instancesState trace logConfig config . handleDelayEffect
-
-handleClient :: InstancesState -> Trace IO LM.PABLogMsg -> CM.Configuration -> Config -> Connection -> IO ()
-handleClient instancesState trace logConfig config connection =
-    handle disconnect . withPingThread connection 30 (pure ()) $
-    threadApp instancesState trace logConfig config connection
+contractInstanceUpdates :: forall t env. ContractInstanceId -> PendingConnection -> PABAction t env ()
+contractInstanceUpdates contractInstanceId pending = do
+    Core.PABRunner{Core.runPABAction} <- Core.pabRunner
+    liftIO $ do
+        connection <- WS.acceptRequest pending
+        handle disconnect . WS.withPingThread connection 30 (pure ()) $ fmap (either (error . show) id) . runPABAction $ sendUpdatesToClient contractInstanceId connection
   where
     disconnect :: SomeException -> IO ()
     disconnect _ = pure ()
 
-handleWS ::
-       ( LastMember m effs
-       , MonadIO m
-       , Member (LogMsg WebSocketLogMsg) effs
-       , Member (Reader Config) effs
-       , Member WebSocketEffect effs
-       )
-    => InstancesState
-    -> Trace IO LM.PABLogMsg
-    -> CM.Configuration
-    -> PendingConnection
-    -> Eff effs ()
-handleWS instancesState trace logConfig pending = do
-    (uuid, connection) <- acceptConnection pending
-    config <- ask
-    logInfo $ CreatedConnection uuid
-    liftIO $ handleClient instancesState trace logConfig config connection
-    logInfo $ ClosedConnection uuid
+getContractReport :: forall t env. Contract.PABContract t => PABAction t env (ContractReport (Contract.ContractDef t))
+getContractReport = do
+    installedContracts <- Contract.getDefinitions @t
+    activeContractIDs <- fmap fst . Map.toList <$> Contract.getActiveContracts @t
+    crAvailableContracts <-
+        traverse
+            (\t -> ContractSignatureResponse t <$> Contract.exportSchema @t t)
+            installedContracts
+    crActiveContractStates <- traverse (\i -> Contract.getState @t i >>= \s -> pure (i, Contract.serialisableState (Proxy @t) s)) activeContractIDs
+    pure ContractReport {crAvailableContracts, crActiveContractStates}
+
+sendUpdatesToClient :: forall t env. ContractInstanceId -> Connection -> PABAction t env ()
+sendUpdatesToClient instanceId connection = do
+    (getState, getEndpoints, finalValue) <- (,,) <$> Core.observableState instanceId <*> fmap (fmap (fmap oepName)) (Core.activeEndpoints instanceId) <*> Core.finalResult instanceId
+    (initialState, initialEndpoints) <- liftIO $ STM.atomically $ (,) <$> getState <*> getEndpoints
+    let nextState oldState = do
+            newState <- getState
+            guard $ newState /= oldState
+            pure newState
+
+        nextEndpoints oldEndpoints = do
+            newEndpoints <- getEndpoints
+            guard $ newEndpoints /= oldEndpoints
+            pure newEndpoints
+
+        go currentState currentEndpoints = do
+            update' <- liftIO $ STM.atomically $ (NewObservableState <$> nextState currentState) <|> (NewActiveEndpoints <$> nextEndpoints currentEndpoints) <|> (ContractFinished <$> finalValue)
+            liftIO $ WS.sendTextData connection $ JSON.encode update'
+            case update' of
+                NewObservableState newState -> go newState currentEndpoints
+                NewActiveEndpoints eps      -> go currentState eps
+                ContractFinished _          -> pure ()
+    go initialState initialEndpoints
